==> 2. Decision Tree Regressor Model.ipynb <==
%load_ext autoreload
%autoreload 2

import numpy as np
import pandas as pd

import time
import json
import requests
from datetime import datetime

pd.set_option("display.max_rows", None)
pd.set_option("display.max_columns", 100)
pd.options.display.float_format = '{:.2f}'.format
np.set_printoptions(formatter={'float': lambda x: "{0:0.2f}".format(x)})

#I. PIPELINE

## accessing the open energy data from Energinet Denmark, through their public API
def get_PowerSystem_data():

    url = 'https://api.energidataservice.dk/dataset/PowerSystemRightNow?start=now-P1Y&end=now&sort=Minutes1DK'
    response = requests.get(url)

    if response.status_code != 200:
        print("Request failed with status code:", response.status_code)

    selected_records = response.json()['records'][::5]
    df = pd.DataFrame(selected_records).fillna(0)

    total_generation = df.ProductionGe100MW + df.ProductionLt100MW + df.SolarPower + df.OffshoreWindPower + df.OnshoreWindPower

    net_imports = df.Exchange_DK1_DE + df.Exchange_DK1_NL + df.Exchange_DK1_GB + df.Exchange_DK1_NO + df.Exchange_DK1_SE + \
                    df.Exchange_DK1_DK2 + df.Exchange_DK2_DE + df.Exchange_DK2_SE + df.Exchange_Bornholm_SE

    imbalance = df.ImbalanceDK1 + df.ImbalanceDK2

    df['Demand'] = total_generation + net_imports - imbalance
    df['Renewables'] = df.SolarPower + df.OffshoreWindPower + df.OnshoreWindPower
    actionable = df['Demand'] - df['Renewables']

    return df

data = get_PowerSystem_data()
data.tail(5)

#II. PREDICT CO2 EMISSION LEVELS BASED ON POWER PRODUCTION

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.tree import DecisionTreeRegressor
from sklearn.metrics import mean_absolute_error


def infer_co2_levels(data):

    X = data.iloc[:, 3:-2]
    X1 = pd.concat([data['Renewables'],X.iloc[:, :2]], axis=1)
    X2 = X.iloc[:, 6:-6]
    X = pd.concat([X1,X2], axis=1).to_numpy()

    y = data.CO2Emission.to_numpy()

    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=9)

    rg = DecisionTreeRegressor()

    param_grid = {
    'max_depth': [None,7,8,9,10,12],
    'max_leaf_nodes': [None,40,45,50]
    }

    grid_search = GridSearchCV(
        estimator=rg,
        param_grid=param_grid,
        scoring='neg_mean_absolute_error',  
        cv=5,                               
        n_jobs=-1,                          
        verbose=1)

    grid_search.fit(X_train, y_train)

    best_params = grid_search.best_params_
    best_score = -grid_search.best_score_
    print("Decision Tree results:")
    print("Best Parameters:", best_params)
    print("Best Mean Absolute Error (MAE) from Cross-Validation:", best_score)

    best_rg = grid_search.best_estimator_
    y_pred = best_rg.predict(X_test)
    test_mae = mean_absolute_error(y_test, y_pred)
    print("Test Mean Absolute Error (MAE):", test_mae)
    print("*"*20)
    print("*"*20)

    return best_rg


from joblib import dump
dump(infer_co2_levels(data), 'decision_tree_regressor.joblib')
print("Model saved successfully.")
