==> 4. XGBoost Renewables Forecast Model.ipynb <==
%load_ext autoreload
%autoreload 2

import numpy as np
import pandas as pd

import time
import json
import requests
from datetime import datetime

pd.set_option("display.max_rows", None)
pd.set_option("display.max_columns", 100)
pd.options.display.float_format = '{:.2f}'.format
np.set_printoptions(formatter={'float': lambda x: "{0:0.2f}".format(x)})

#I. PIPELINE

## accessing the open energy data from Energinet Denmark, through their public API
def get_PowerSystem_data():

    url = 'https://api.energidataservice.dk/dataset/PowerSystemRightNow?start=now-P1Y&end=now&sort=Minutes1DK'
    response = requests.get(url)

    if response.status_code != 200:
        print("Request failed with status code:", response.status_code)

    selected_records = response.json()['records'][::5]
    df = pd.DataFrame(selected_records).fillna(0)

    total_generation = df.ProductionGe100MW + df.ProductionLt100MW + df.SolarPower + df.OffshoreWindPower + df.OnshoreWindPower

    net_imports = df.Exchange_DK1_DE + df.Exchange_DK1_NL + df.Exchange_DK1_GB + df.Exchange_DK1_NO + df.Exchange_DK1_SE + \
                    df.Exchange_DK1_DK2 + df.Exchange_DK2_DE + df.Exchange_DK2_SE + df.Exchange_Bornholm_SE

    imbalance = df.ImbalanceDK1 + df.ImbalanceDK2

    df['Demand'] = total_generation + net_imports - imbalance
    df['Renewables'] = df.SolarPower + df.OffshoreWindPower + df.OnshoreWindPower
    actionable = df['Demand'] - df['Renewables']

    return df

data = get_PowerSystem_data()
data.tail(5)

#II. PREDICT RENEWABLE ENERGY PRODUCTION

##demand forecast

from xgboost import XGBRegressor
from sklearn.metrics import root_mean_squared_error
from sklearn.model_selection import train_test_split, GridSearchCV

def predict_renewables(data):

    series = data.Renewables

    window_size = 10
    n_steps = 5

    ## function to create features and labels
    def create_multi_step_features(data, window_size, n_steps):
        X, y = [], []
        for i in range(len(data) - window_size - n_steps):
            X.append(data[i:i + window_size])
            y.append(data[i + window_size:i + window_size + n_steps])
        return np.array(X), np.array(y)

    ## create the features and labels
    X, y = create_multi_step_features(series, window_size, n_steps)

    train_size = int(len(X) * 0.8)
    X_train, X_test = X[:train_size], X[train_size:]
    y_train, y_test = y[:train_size], y[train_size:]

    param_grid = {
        'n_estimators': [50, 100, 150],
        'learning_rate': [0.01, 0.1, 0.2],
        'max_depth': [4, 6, 8],
        'subsample': [0.8, 1.0],
        'colsample_bytree': [0.8, 1.0]
    }

    print("Renewables Forecast XGBoost results:")

    best_xgb_models = []
    for step in range(n_steps):
        xgb_model = XGBRegressor(objective='reg:squarederror')
        grid_search = GridSearchCV(estimator=xgb_model, param_grid=param_grid, 
                                   scoring='neg_mean_squared_error', verbose=1, n_jobs=-1)
        
        grid_search.fit(X_train, y_train[:, step])

        best_xgb_model = grid_search.best_estimator_
        best_xgb_models.append(best_xgb_model)

        print(f"Best parameters for step {step+1}: {grid_search.best_params_}")

    y_pred = []
    for best_xgb_model in best_xgb_models:
        y_pred.append(best_xgb_model.predict(X_test))

    y_pred = np.array(y_pred).T

    for step in range(n_steps):
        mse = root_mean_squared_error(y_test[:, step], y_pred[:, step])
        print(f"Mean Squared Error for step {step+1}: {mse}")

    print(" "*20)
    print("*"*20)

    return best_xgb_models

from joblib import dump
dump(predict_renewables(data), 'XGBoost_regressor_renewables.joblib')
print("Model saved successfully.")
